todolist file refactored to different modules
