+++ src/index.js
@@ -1,10 +1,10 @@
 import '@fortawesome/fontawesome-free/css/all.css';
 import '@fortawesome/fontawesome-free/js/all.js';
 import './style.css';
-import Todo from './modules/todoList.js';
 import Completed from './modules/completed.js';
+import addList from './modules/addTask.js';
+import showList from './modules/showTasks.js';
 
-const todoClass = new Todo();
 const completed = new Completed();
 
 const addButton = document.querySelector('.addList');
@@ -13,11 +13,11 @@ const clearButton = document.querySelector('.clear');
 addButton.addEventListener('click', (event) => {
   event.preventDefault();
   const inputValue = document.querySelector('.listInput').value;
-  if (inputValue !== '') todoClass.addList();
+  if (inputValue !== '') addList(inputValue);
 });
 
 clearButton.addEventListener('click', () => {
   completed.clearCompleted();
 });
 
-todoClass.showList();
+showList();
