+++ src/modules/completed.js
@@ -1,8 +1,7 @@
-import Todo from './todoList.js';
+import showList from './showTasks.js';
 
 export default class Completed {
     clearCompleted = () => {
-      const todo = new Todo();
       const listCollection = JSON.parse(localStorage.getItem('Lists'));
       const listUpdated = listCollection
         .filter((list) => !list.completed)
@@ -11,6 +10,6 @@ export default class Completed {
           index: index + 1,
         }));
       localStorage.setItem('Lists', JSON.stringify(listUpdated));
-      todo.showList();
+      showList();
     };
 }
